backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser==2.2.0
Pillow==9.3.0
psycopg2-binary==2.9.5
django-filter==21.1
gunicorn==20.1.0
